Fix ignoredCancelled to be clearer
Also fixes the #of ignoring canceled events by default; which would be the opposite behavior of the builder.

This also fixes the cancelable test and adds new ones to ensure correctness in the future of its behavior.

This is a breaking change for anyone using the inital broken assumption as its been backwords for